Update:
Somewhere the code became wonky and we had to debug.  Ive fixed the code so that it would no longer throw the errors when migration process was run...

			Warning: 1265 Data truncated for column

The error we received was the warning listed above. In order to combat this we changed the INT assignment to a DECIMAL assignment.