Update:
Seeder wasn't allowing for data to be injected into the database with the hardcoded "Listings" we created.  We received an error and found out we had left a couple of semi-colons off of our descriptions.

We fixed the bug by adding the semi_colons and re-runing our seeder.  Everything checked out okay and our data is being added accordingly.