Update:
Had a bug that wouldn't allow the injection of data via the seeder file. We've since corrected the file and data can now be injected via the seeder.  Additionally, we created an item_image field that will allow users the option to upload a an item image.  If you have any input on items you'd like to see in the database, let me know.